requirements.txt
Pillow==4.0.0
opencv-python==3.2.0.6
numpy==1.12.0
PyYAML==3.12
pypiwin32==219
requests==2.11.1
